FACE_ATTENDENCE/requirements.txt
opencv-python
face-recognition
numpy
pandas
python-dateutil
flask
pillow
streamlit